test(piece): address retry review feedback
